+++ src/Provider/PaymentOption/BancontactPaymentOptionProvider.php
@@ -48,6 +48,8 @@ use Tools;
 
 class BancontactPaymentOptionProvider implements PaymentOptionProviderInterface
 {
+    const FILE_NAME = 'BancontactPaymentOptionProvider';
+
     /**
      * @var Mollie
      */
@@ -135,7 +137,7 @@ class BancontactPaymentOptionProvider implements PaymentOptionProviderInterface
                     [
                         'type' => 'hidden',
                         'name' => 'payment-fee-price-display',
-                        'value' => sprintf($this->module->l('Payment Fee: %1s'), Tools::displayPrice($paymentFee)),
+                        'value' => sprintf($this->module->l('Payment Fee: %1s', self::FILE_NAME), Tools::displayPrice($paymentFee)),
                     ],
                     [
                         'type' => 'hidden',