+++ src/Provider/PaymentOption/IdealPaymentOptionProvider.php
@@ -49,6 +49,8 @@ use Tools;
 
 class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
 {
+    const FILE_NAME = 'IdealPaymentOptionProvider';
+
     /**
      * @var Mollie
      */
@@ -152,7 +154,7 @@ class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
                     [
                         'type' => 'hidden',
                         'name' => 'payment-fee-price-display',
-                        'value' => sprintf($this->module->l('Payment Fee: %1s'), Tools::displayPrice($paymentFee)),
+                        'value' => sprintf($this->module->l('Payment Fee: %1s', self::FILE_NAME), Tools::displayPrice($paymentFee)),
                     ],
                 ]
             );