+++ src/Provider/ProfileIdProviderInterface.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Mollie       https://www.mollie.nl
+ *
+ * @author      Mollie B.V. <info@mollie.nl>
+ * @copyright   Mollie B.V.
+ * @license     https://github.com/mollie/PrestaShop/blob/master/LICENSE.md
+ *
+ * @see        https://github.com/mollie/PrestaShop
+ * @codingStandardsIgnoreStart
+ */
+
+namespace Mollie\Provider;
+
+use Mollie\Api\MollieApiClient;
+
+interface ProfileIdProviderInterface
+{
+    public function getProfileId(MollieApiClient $apiClient): string;
+}