+++ tests/Unit/Tools/UnitTestCase.php
@@ -1,4 +1,14 @@
 <?php
+/**
+ * Mollie       https://www.mollie.nl
+ *
+ * @author      Mollie B.V. <info@mollie.nl>
+ * @copyright   Mollie B.V.
+ * @license     https://github.com/mollie/PrestaShop/blob/master/LICENSE.md
+ *
+ * @see        https://github.com/mollie/PrestaShop
+ * @codingStandardsIgnoreStart
+ */
 
 namespace Mollie\Tests\Unit\Tools;
 
@@ -51,6 +61,11 @@ class UnitTestCase extends TestCase
 
         $contextMock
             ->method('getCountryId')
+            ->willReturn(1)
+        ;
+
+        $contextMock
+            ->method('getInvoiceCountryId')
             ->willReturn(1)
         ;